refactor(web): courts-stat
